Refactor - rename `apiHelper`
